catalog: carve FlowRunService out of the facade
Move FlowRun lifecycle (list, get, start, complete, create_completed,
snapshot, run_flow_now, list_active, cancel) plus the spawn helpers
and _resolve_log_path into catalog/services/runs.py. The facade keeps
thin forwarders so the public API stays byte-identical.

run_to_out (the impure version that needs FS access) becomes a
public method on FlowRunService; pure run_to_out from serializers
takes has_log explicitly.
