+++ flowfile_core/flowfile_core/catalog/service.py
@@ -69,6 +69,7 @@ from flowfile_core.catalog.exceptions import (
 )
 from flowfile_core.catalog.repository import CatalogRepository
 from flowfile_core.catalog.services.namespaces import NamespaceService
+from flowfile_core.catalog.services.runs import FlowRunService
 from flowfile_core.catalog.serializers import (
     VizEnrichment,
     artifact_to_out,
@@ -175,6 +176,7 @@ class CatalogService:
     def __init__(self, repo: CatalogRepository) -> None:
         self.repo = repo
         self._namespaces = NamespaceService(repo)
+        self._runs = FlowRunService(repo)
 
     # ------------------------------------------------------------------ #
     # Private helpers
@@ -446,18 +448,11 @@ class CatalogService:
             )
         return result
 
-    @staticmethod
-    def _resolve_log_path(run_id: int, run_type: str) -> str | None:
-        """Return the log file path if it exists for subprocess-spawned runs."""
-        if run_type not in ("scheduled", "manual", "on_demand"):
-            return None
-        log_file = Path.home() / ".flowfile" / "logs" / f"scheduled_run_{run_id}.log"
-        if log_file.exists():
-            return str(log_file)
-        return None
+    def _resolve_log_path(self, run_id: int, run_type: str) -> str | None:
+        return self._runs._resolve_log_path(run_id, run_type)
 
     def _run_to_out(self, run: FlowRun) -> FlowRunOut:
-        return run_to_out(run, has_log=self._resolve_log_path(run.id, run.run_type) is not None)
+        return self._runs.run_to_out(run)
 
     # ------------------------------------------------------------------ #
     # Namespace operations
@@ -711,64 +706,13 @@ class CatalogService:
         limit: int = 50,
         offset: int = 0,
     ) -> PaginatedFlowRuns:
-        """List run summaries (without snapshots) with total count for pagination."""
-        runs = self.repo.list_runs(
-            registration_id=registration_id, schedule_id=schedule_id, run_type=run_type, limit=limit, offset=offset
-        )
-        counts = self.repo.count_runs_by_status(
-            registration_id=registration_id, schedule_id=schedule_id, run_type=run_type
-        )
-        return PaginatedFlowRuns(
-            items=[self._run_to_out(r) for r in runs],
-            total=counts["total"],
-            total_success=counts["success"],
-            total_failed=counts["failed"],
-            total_running=counts["running"],
-        )
+        return self._runs.list_runs(registration_id, schedule_id, run_type, limit, offset)
 
     def get_run_detail(self, run_id: int) -> FlowRunDetail:
-        """Get a single run including the YAML snapshot.
-
-        Raises
-        ------
-        RunNotFoundError
-            If the run doesn't exist.
-        """
-        run = self.repo.get_run(run_id)
-        if run is None:
-            raise RunNotFoundError(run_id=run_id)
-        return FlowRunDetail(
-            id=run.id,
-            registration_id=run.registration_id,
-            flow_name=run.flow_name,
-            flow_path=run.flow_path,
-            user_id=run.user_id,
-            started_at=run.started_at,
-            ended_at=run.ended_at,
-            success=run.success,
-            nodes_completed=run.nodes_completed,
-            number_of_nodes=run.number_of_nodes,
-            duration_seconds=run.duration_seconds,
-            run_type=run.run_type,
-            schedule_id=run.schedule_id,
-            has_snapshot=run.flow_snapshot is not None,
-            has_log=self._resolve_log_path(run.id, run.run_type) is not None,
-            flow_snapshot=run.flow_snapshot,
-            node_results_json=run.node_results_json,
-        )
+        return self._runs.get_run_detail(run_id)
 
     def get_run(self, run_id: int) -> FlowRun:
-        """Get a raw FlowRun model.
-
-        Raises
-        ------
-        RunNotFoundError
-            If the run doesn't exist.
-        """
-        run = self.repo.get_run(run_id)
-        if run is None:
-            raise RunNotFoundError(run_id=run_id)
-        return run
+        return self._runs.get_run(run_id)
 
     def start_run(
         self,
@@ -780,18 +724,9 @@ class CatalogService:
         run_type: RunType,
         flow_snapshot: str | None = None,
     ) -> FlowRun:
-        """Record a new flow run start."""
-        run = FlowRun(
-            registration_id=registration_id,
-            flow_name=flow_name,
-            flow_path=flow_path,
-            user_id=user_id,
-            started_at=datetime.now(timezone.utc),
-            number_of_nodes=number_of_nodes,
-            run_type=run_type,
-            flow_snapshot=flow_snapshot,
+        return self._runs.start_run(
+            registration_id, flow_name, flow_path, user_id, number_of_nodes, run_type, flow_snapshot
         )
-        return self.repo.create_run(run)
 
     def complete_run(
         self,
@@ -801,31 +736,7 @@ class CatalogService:
         number_of_nodes: int | None = None,
         node_results_json: str | None = None,
     ) -> FlowRun:
-        """Mark a run as completed.
-
-        Raises
-        ------
-        RunNotFoundError
-            If the run doesn't exist.
-        """
-        run = self.repo.get_run(run_id)
-        if run is None:
-            raise RunNotFoundError(run_id=run_id)
-        now = datetime.now(timezone.utc)
-        run.ended_at = now
-        run.success = success
-        run.nodes_completed = nodes_completed
-        if number_of_nodes is not None and number_of_nodes > 0:
-            run.number_of_nodes = number_of_nodes
-        if run.started_at:
-            # Normalize both sides to naive UTC for duration calculation.
-            # SQLite stores naive datetimes, so started_at may lack tzinfo.
-            started_utc = run.started_at.replace(tzinfo=None)
-            now_utc = now.replace(tzinfo=None)
-            run.duration_seconds = (now_utc - started_utc).total_seconds()
-        if node_results_json is not None:
-            run.node_results_json = node_results_json
-        return self.repo.update_run(run)
+        return self._runs.complete_run(run_id, success, nodes_completed, number_of_nodes, node_results_json)
 
     def create_completed_run(
         self,
@@ -842,26 +753,20 @@ class CatalogService:
         node_results_json: str | None = None,
         flow_snapshot: str | None = None,
     ) -> FlowRun:
-        """Record a fully completed run in one step (fallback when start_run was skipped)."""
-        duration = None
-        if started_at and ended_at:
-            duration = (ended_at - started_at).total_seconds()
-        run = FlowRun(
-            registration_id=registration_id,
-            flow_name=flow_name,
-            flow_path=flow_path,
-            user_id=user_id,
-            started_at=started_at,
-            ended_at=ended_at,
-            success=success,
-            nodes_completed=nodes_completed,
-            number_of_nodes=number_of_nodes,
-            duration_seconds=duration,
-            run_type=run_type,
-            node_results_json=node_results_json,
-            flow_snapshot=flow_snapshot,
+        return self._runs.create_completed_run(
+            registration_id,
+            flow_name,
+            flow_path,
+            user_id,
+            started_at,
+            ended_at,
+            success,
+            nodes_completed,
+            number_of_nodes,
+            run_type,
+            node_results_json,
+            flow_snapshot,
         )
-        return self.repo.create_run(run)
 
     def auto_register_flow(self, flow_path: str, name: str, user_id: int) -> FlowRegistration | None:
         """Register a flow in a default-ish namespace if not already registered.
@@ -940,21 +845,7 @@ class CatalogService:
         return reg.id if reg else None
 
     def get_run_snapshot(self, run_id: int) -> str:
-        """Return the flow snapshot text for a run.
-
-        Raises
-        ------
-        RunNotFoundError
-            If the run doesn't exist.
-        NoSnapshotError
-            If the run has no snapshot.
-        """
-        run = self.repo.get_run(run_id)
-        if run is None:
-            raise RunNotFoundError(run_id=run_id)
-        if not run.flow_snapshot:
-            raise NoSnapshotError(run_id=run_id)
-        return run.flow_snapshot
+        return self._runs.get_run_snapshot(run_id)
 
     # ------------------------------------------------------------------ #
     # Favorites
@@ -2481,14 +2372,8 @@ class CatalogService:
     # Trigger schedule now
     # ------------------------------------------------------------------ #
 
-    @staticmethod
-    def _spawn_flow_subprocess(flow_path: str, run_id: int) -> int | None:
-        """Fire-and-forget a ``flowfile run flow`` subprocess.
-
-        Returns the child PID on success, or ``None`` on failure.
-        """
-
-        return spawn_flow_subprocess(flow_path, run_id)
+    def _spawn_flow_subprocess(self, flow_path: str, run_id: int) -> int | None:
+        return self._runs._spawn_flow_subprocess(flow_path, run_id)
 
     def _spawn_flow_run(
         self,
@@ -2497,52 +2382,10 @@ class CatalogService:
         run_type: RunType,
         schedule_id: int | None = None,
     ) -> FlowRun:
-        """Create a FlowRun record and spawn the subprocess.
-
-        On spawn failure the run is marked as failed immediately.
-        """
-        now = datetime.now(timezone.utc)
-        run = FlowRun(
-            registration_id=flow.id,
-            flow_name=flow.name,
-            flow_path=flow.flow_path,
-            user_id=user_id,
-            started_at=now,
-            number_of_nodes=0,
-            run_type=run_type,
-            schedule_id=schedule_id,
-        )
-        run = self.repo.create_run(run)
-
-        pid = self._spawn_flow_subprocess(flow.flow_path, run.id)
-        if pid is not None:
-            run.pid = pid
-        else:
-            logger.error("Failed to spawn subprocess for run %s — marking as failed", run.id)
-            run.ended_at = datetime.now(timezone.utc)
-            run.success = False
-        self.repo.update_run(run)
-        return run
+        return self._runs.spawn_flow_run(flow, user_id, run_type, schedule_id)
 
     def run_flow_now(self, registration_id: int, user_id: int) -> FlowRunOut:
-        """Trigger a registered flow immediately without a schedule.
-
-        Raises
-        ------
-        FlowNotFoundError
-            If the flow doesn't exist.
-        FlowAlreadyRunningError
-            If the flow already has an active (unfinished) run.
-        """
-        flow = self.repo.get_flow(registration_id)
-        if flow is None:
-            raise FlowNotFoundError(registration_id=registration_id)
-
-        if self.repo.has_active_run(registration_id):
-            raise FlowAlreadyRunningError(registration_id=registration_id)
-
-        run = self._spawn_flow_run(flow, user_id=user_id, run_type="manual")
-        return self._run_to_out(run)
+        return self._runs.run_flow_now(registration_id, user_id)
 
     def trigger_schedule_now(self, schedule_id: int, user_id: int) -> FlowRunOut:
         """Manually trigger a scheduled flow immediately.
@@ -2576,58 +2419,10 @@ class CatalogService:
     # ------------------------------------------------------------------ #
 
     def list_active_runs(self) -> list[ActiveFlowRun]:
-        """List all currently running flows (ended_at IS NULL)."""
-        runs = self.repo.list_active_runs()
-        return [
-            ActiveFlowRun(
-                id=r.id,
-                registration_id=r.registration_id,
-                flow_name=r.flow_name,
-                flow_path=r.flow_path,
-                user_id=r.user_id,
-                started_at=r.started_at,
-                nodes_completed=r.nodes_completed,
-                number_of_nodes=r.number_of_nodes,
-                run_type=r.run_type,
-            )
-            for r in runs
-        ]
+        return self._runs.list_active_runs()
 
     def cancel_run(self, run_id: int) -> None:
-        """Cancel a running flow by terminating its subprocess and marking
-        the database record as failed.
-
-        If ``pid`` is stored on the run, sends ``SIGTERM`` to the process.
-        ``ProcessLookupError`` is silently ignored (the process already
-        exited).  If no PID is available the record is still marked as
-        cancelled.
-
-        Raises
-        ------
-        RunNotFoundError
-            If the run doesn't exist.
-        """
-        run = self.repo.get_run(run_id)
-        if run is None:
-            raise RunNotFoundError(run_id=run_id)
-
-        if run.pid is not None:
-            try:
-                os.kill(run.pid, signal.SIGTERM)
-                logger.info("Sent SIGTERM to pid %s for run %s", run.pid, run_id)
-            except ProcessLookupError:
-                logger.info("Process %s for run %s already exited", run.pid, run_id)
-            except OSError:
-                logger.warning("Failed to kill pid %s for run %s", run.pid, run_id, exc_info=True)
-
-        now = datetime.now(timezone.utc)
-        run.ended_at = now
-        run.success = False
-        if run.started_at:
-            started_utc = run.started_at.replace(tzinfo=None)
-            now_utc = now.replace(tzinfo=None)
-            run.duration_seconds = (now_utc - started_utc).total_seconds()
-        self.repo.update_run(run)
+        self._runs.cancel_run(run_id)
 
     # ------------------------------------------------------------------ #
     # Dashboard / Stats
